checkloc, tests - still run if passed exact path to chrome.manifest
Create a more forgiving UI by still attempting to run
if the user passes in the exact path to the chrome.manifest file
rather than the parent directory.
It would be silly to reject them and throw an error stating
to only pass the directory
when we can easily figure out the correct path ourselves.